sha256 into localrepo, chksum only 1 method
