+++ tests/unit/test_localrepository.py
@@ -11,6 +11,11 @@ try:
 except ImportError:
     from mock import Mock, patch, MagicMock, mock_open, call
 
+try:
+    from StringIO import StringIO
+except ImportError:
+    from io import StringIO
+
 sys.path.append('.')
 
 from udocker.container.localrepo import LocalRepository
@@ -66,6 +71,24 @@ class LocalRepositoryTestCase(TestCase):
         lrepo = LocalRepository(self.conf)
         lrepo.setup("YYYY")
         self.assertEqual(os.path.basename(lrepo.topdir), "YYYY")
+
+    def test_01_sha256(self):
+        """Test LocalRepository().sha256()."""
+        sha256sum = (
+            "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855")
+        lrepo = LocalRepository(self.conf)
+        file_data = StringIO("qwerty")
+        with patch(BOPEN, mock_open()) as mopen:
+            mopen.return_value.__iter__ = (
+                lambda self: iter(file_data.readline, ''))
+            status = lrepo.sha256("filename")
+            self.assertEqual(status, sha256sum)
+
+        with patch(BOPEN, mock_open()) as mopen:
+            mopen.side_effect = IOError
+            status = lrepo.sha256("filename")
+            self.assertRaises(IOError)
+            self.assertEqual(status, "")
 
     def test_03_create_repo(self):
         """Test LocalRepository().create_repo()."""
